added sync sqs client
