+++ requirements.txt
@@ -1,4 +1,6 @@
 cuenca-validations==2.1.0
+boto3==1.35.74
+types-boto3[sqs]==1.35.74
 chalice==1.31.3
 mongoengine==0.29.1
 fastapi==0.115.6
