+++ setup.py
@@ -39,6 +39,14 @@ setup(
             'aiobotocore>=2.0.0,<3.0.0',
             'types-aiobotocore-sqs>=2.1.0,<3.0.0',
         ],
+        'sync_aws_tools': [
+            'boto3>=1.34.106,<2.0.0',
+            'types-boto3[sqs]>=1.34.106,<2.0.0',
+        ],
+        'async_aws_tools': [
+            'aiobotocore>=2.0.0,<3.0.0',
+            'types-aiobotocore-sqs>=2.1.0,<3.0.0',
+        ],
     },
     classifiers=[
         'Programming Language :: Python :: 3.9',
