+++ tests/conftest.py
@@ -1,13 +1,20 @@
 import datetime as dt
 import functools
 import json
+import os
+from functools import partial
 from typing import Callable, Generator
 
+import aiobotocore
+import boto3
 import pytest
+from _pytest.monkeypatch import MonkeyPatch
+from aiobotocore.session import AioSession
 from chalice.test import Client as OriginalChaliceClient
 from fastapi.testclient import TestClient as FastAPIClient
 from mongoengine import Document
 
+from agave.tasks import sqs_tasks
 from examples.config import (
     TEST_DEFAULT_PLATFORM_ID,
     TEST_DEFAULT_USER_ID,
@@ -216,3 +223,93 @@ def chalice_client() -> Generator[ChaliceClient, None, None]:
 
     client = ChaliceClient(app)
     yield client
+
+
+@pytest.fixture(scope='session')
+def aws_credentials() -> None:
+    """Mocked AWS Credentials for moto."""
+    os.environ['AWS_ACCESS_KEY_ID'] = 'testing'
+    os.environ['AWS_SECRET_ACCESS_KEY'] = 'testing'
+    os.environ['AWS_SECURITY_TOKEN'] = 'testing'
+    os.environ['AWS_DEFAULT_REGION'] = 'us-east-1'
+    boto3.setup_default_session()
+
+
+@pytest.fixture(scope='session')
+def aws_endpoint_urls(
+    aws_credentials,
+) -> Generator[dict[str, str], None, None]:
+    from moto.server import ThreadedMotoServer
+
+    server = ThreadedMotoServer(port=4000)
+    server.start()
+
+    endpoints = dict(
+        sqs='http://127.0.0.1:4000/',
+    )
+    yield endpoints
+
+    server.stop()
+
+
+@pytest.fixture(autouse=True)
+def patch_tasks_count(monkeypatch: MonkeyPatch) -> None:
+    def one_loop(*_, **__):
+        # Para pruebas solo unos cuantos ciclos
+        for i in range(5):
+            yield i
+
+    monkeypatch.setattr(sqs_tasks, 'count', one_loop)
+
+
+@pytest.fixture(autouse=True)
+def patch_aiobotocore_create_client(
+    aws_endpoint_urls, monkeypatch: MonkeyPatch
+) -> None:
+    create_client = AioSession.create_client
+
+    def mock_create_client(*args, **kwargs):
+        service_name = next(a for a in args if type(a) is str)
+        kwargs['endpoint_url'] = aws_endpoint_urls[service_name]
+
+        return create_client(*args, **kwargs)
+
+    monkeypatch.setattr(AioSession, 'create_client', mock_create_client)
+
+
+@pytest.fixture(autouse=True)
+def patch_boto3_create_client(
+    aws_endpoint_urls, monkeypatch: MonkeyPatch
+) -> None:
+    create_client = boto3.Session.client
+
+    def mock_client(*args, **kwargs):
+        service_name = next(a for a in args if type(a) is str)
+        if service_name in aws_endpoint_urls:
+            kwargs['endpoint_url'] = aws_endpoint_urls[service_name]
+        return create_client(*args, **kwargs)
+
+    monkeypatch.setattr(boto3.Session, 'client', mock_client)
+
+
+@pytest.fixture
+async def sqs_client():
+    session = aiobotocore.session.get_session()
+    async with session.create_client('sqs', 'us-east-1') as sqs:
+        await sqs.create_queue(
+            QueueName='core.fifo',
+            Attributes={
+                'FifoQueue': 'true',
+                'ContentBasedDeduplication': 'true',
+            },
+        )
+        resp = await sqs.get_queue_url(QueueName='core.fifo')
+        sqs.send_message = partial(sqs.send_message, QueueUrl=resp['QueueUrl'])
+        sqs.receive_message = partial(
+            sqs.receive_message,
+            QueueUrl=resp['QueueUrl'],
+            AttributeNames=['ApproximateReceiveCount'],
+        )
+        sqs.queue_url = resp['QueueUrl']
+        yield sqs
+        await sqs.purge_queue(QueueUrl=resp['QueueUrl'])
